+++ app/Repositories/CropProductionRepository.php
@@ -14,9 +14,42 @@ class CropProductionRepository implements CropProductionRepositoryInterface
         $this->model = $model;
     }
 
-    public function all()
+    public function all(array $filters = [])
     {
-        return $this->model->with(['region'])->get();
+        $query = $this->model->query()->with(['region']);
+
+        if (isset($filters['region_id'])) {
+            $query->where('region_id', $filters['region_id']);
+        }
+
+        if (isset($filters['crop_name'])) {
+            $query->where('crop_name', 'like', '%' . $filters['crop_name'] . '%');
+        }
+
+        if (isset($filters['fiscal_year'])) {
+            $query->where('fiscal_year', $filters['fiscal_year']);
+        }
+
+        return $query->get();
+    }
+
+    public function paginate(int $perPage = 15, array $filters = [])
+    {
+        $query = $this->model->query()->with(['region']);
+
+        if (isset($filters['region_id'])) {
+            $query->where('region_id', $filters['region_id']);
+        }
+
+        if (isset($filters['crop_name'])) {
+            $query->where('crop_name', 'like', '%' . $filters['crop_name'] . '%');
+        }
+
+        if (isset($filters['fiscal_year'])) {
+            $query->where('fiscal_year', $filters['fiscal_year']);
+        }
+
+        return $query->paginate($perPage);
     }
 
     public function find($id)
@@ -32,23 +65,29 @@ class CropProductionRepository implements CropProductionRepositoryInterface
     public function update($id, array $data)
     {
         $cropProduction = $this->find($id);
-        $cropProduction->update($data);
-        return $cropProduction;
+        if ($cropProduction) {
+            $cropProduction->update($data);
+            return $cropProduction->fresh();
+        }
+        return null;
     }
 
     public function delete($id)
     {
         $cropProduction = $this->find($id);
-        $cropProduction->delete();
-        return $cropProduction;
+        if ($cropProduction) {
+            $cropProduction->delete();
+            return $cropProduction;
+        }
+        return null;
     }
 
     public function getByRegion($regionId)
     {
-        return $this->model->where('region_id', $regionId)->get();
+        return $this->model->where('region_id', $regionId)->with(['region'])->get();
     }
 
-    public function getByFiscalYear($fiscalYear)
+    public function getByFiscalYear(int $fiscalYear)
     {
         return $this->model->where('fiscal_year', $fiscalYear)
             ->with(['region'])
