+++ app/Repositories/DocumentRepository.php
@@ -14,9 +14,40 @@ class DocumentRepository implements DocumentRepositoryInterface
         $this->model = $model;
     }
 
-    public function all()
+    public function all(array $filters = [])
     {
-        return $this->model->with(['creator', 'categories'])->get();
+        $query = $this->model->query()->with(['creator', 'categories']);
+
+        if (isset($filters['document_type'])) {
+            $query->where('document_type', $filters['document_type']);
+        }
+
+        if (isset($filters['search'])) {
+            $query->where(function ($q) use ($filters) {
+                $q->where('title', 'like', '%' . $filters['search'] . '%')
+                  ->orWhere('description', 'like', '%' . $filters['search'] . '%');
+            });
+        }
+
+        return $query->get();
+    }
+
+    public function paginate(int $perPage = 15, array $filters = [])
+    {
+        $query = $this->model->query()->with(['creator', 'categories']);
+
+        if (isset($filters['document_type'])) {
+            $query->where('document_type', $filters['document_type']);
+        }
+
+        if (isset($filters['search'])) {
+            $query->where(function ($q) use ($filters) {
+                $q->where('title', 'like', '%' . $filters['search'] . '%')
+                  ->orWhere('description', 'like', '%' . $filters['search'] . '%');
+            });
+        }
+
+        return $query->paginate($perPage);
     }
 
     public function find($id)
@@ -32,22 +63,47 @@ class DocumentRepository implements DocumentRepositoryInterface
     public function update($id, array $data)
     {
         $document = $this->find($id);
-        $document->update($data);
-        return $document;
+        if ($document) {
+            $document->update($data);
+            return $document->fresh();
+        }
+        return null;
     }
 
     public function delete($id)
     {
         $document = $this->find($id);
-        $document->delete();
-        return $document;
+        if ($document) {
+            $document->delete();
+            return $document;
+        }
+        return null;
     }
 
-    public function getByType($type)
+    public function search(string $query)
+    {
+        return $this->model->where(function ($q) use ($query) {
+            $q->where('title', 'like', '%' . $query . '%')
+              ->orWhere('description', 'like', '%' . $query . '%')
+              ->orWhere('file_name', 'like', '%' . $query . '%');
+        })->with(['creator', 'categories'])->get();
+    }
+
+    public function getByType(string $type)
     {
         return $this->model->where('document_type', $type)
             ->with(['creator', 'categories'])
             ->get();
+    }
+
+    public function syncCategories($documentId, array $categoryIds)
+    {
+        $document = $this->find($documentId);
+        if ($document) {
+            $document->categories()->sync($categoryIds);
+            return $document->load('categories');
+        }
+        return null;
     }
 
     public function getByCategory($categoryId)
