+++ app/Repositories/LivestockStatisticRepository.php
@@ -14,9 +14,42 @@ class LivestockStatisticRepository implements LivestockStatisticRepositoryInterf
         $this->model = $model;
     }
 
-    public function all()
+    public function all(array $filters = [])
     {
-        return $this->model->with(['region'])->get();
+        $query = $this->model->query()->with(['region']);
+
+        if (isset($filters['region_id'])) {
+            $query->where('region_id', $filters['region_id']);
+        }
+
+        if (isset($filters['livestock_type'])) {
+            $query->where('livestock_type', 'like', '%' . $filters['livestock_type'] . '%');
+        }
+
+        if (isset($filters['fiscal_year'])) {
+            $query->where('fiscal_year', $filters['fiscal_year']);
+        }
+
+        return $query->get();
+    }
+
+    public function paginate(int $perPage = 15, array $filters = [])
+    {
+        $query = $this->model->query()->with(['region']);
+
+        if (isset($filters['region_id'])) {
+            $query->where('region_id', $filters['region_id']);
+        }
+
+        if (isset($filters['livestock_type'])) {
+            $query->where('livestock_type', 'like', '%' . $filters['livestock_type'] . '%');
+        }
+
+        if (isset($filters['fiscal_year'])) {
+            $query->where('fiscal_year', $filters['fiscal_year']);
+        }
+
+        return $query->paginate($perPage);
     }
 
     public function find($id)
@@ -32,23 +65,29 @@ class LivestockStatisticRepository implements LivestockStatisticRepositoryInterf
     public function update($id, array $data)
     {
         $livestockStatistic = $this->find($id);
-        $livestockStatistic->update($data);
-        return $livestockStatistic;
+        if ($livestockStatistic) {
+            $livestockStatistic->update($data);
+            return $livestockStatistic->fresh();
+        }
+        return null;
     }
 
     public function delete($id)
     {
         $livestockStatistic = $this->find($id);
-        $livestockStatistic->delete();
-        return $livestockStatistic;
+        if ($livestockStatistic) {
+            $livestockStatistic->delete();
+            return $livestockStatistic;
+        }
+        return null;
     }
 
     public function getByRegion($regionId)
     {
-        return $this->model->where('region_id', $regionId)->get();
+        return $this->model->where('region_id', $regionId)->with(['region'])->get();
     }
 
-    public function getByFiscalYear($fiscalYear)
+    public function getByFiscalYear(int $fiscalYear)
     {
         return $this->model->where('fiscal_year', $fiscalYear)
             ->with(['region'])
